Backend/NotificationService/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
frozenlist==1.6.0
idna==3.10
kombu==5.5.3
multidict==6.4.3
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
redis==5.2.1
six==1.17.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.20.0
